+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Crash when teleporting on servers [#447]
+
 ## [3.5.0] - 2020-02-03
 Ported to Minecraft 1.15.2
 ### Changed