+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [3.5.1] - 2020-02-19
 ### Added
 - Configs to control number of gem types per biome in the overworld [#448]
 - Config options for geode generation chances [#448]
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Setting world gen region sizes to zero will now disable gem generation in that dimension [#448]
 ### Fixed
 - Crash when teleporting on servers [#447]
+- Pedestal items rendering incorrectly [#443]
 
 ## [3.5.0] - 2020-02-03
 Ported to Minecraft 1.15.2